pyproject.toml
[tool.poetry]
name = "pd-rest-api-contacts"
version = "0.1.0"
description = ""
authors = ["pavlodubovyi <133531274+pavlodubovyi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
databases = "^0.9.0"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
jose = "^1.0.0"
fastapi-jwt-auth = "^0.5.0"
python-multipart = "^0.0.17"
fastapi-mail = "^1.4.1"
cloudinary = "^1.41.0"
itsdangerous = "^2.2.0"
fastapi-limiter = "^0.1.6"
redis = "^5.2.0"
bcrypt = "^4.2.0"
black = "^24.10.0"
sphinx = "^8.1.3"
starlette = "^0.41.2"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
